% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books_and_serials_sample_data.R
\docType{data}
\name{books_and_serials_sample}
\alias{books_and_serials_sample}
\title{Small sample of books, monographs, and serials and their information}
\format{
An object of class \code{"data.table"};
}
\usage{
data(books_and_serials_sample)
}
\description{
A very small sample of books, monographs, and serials and their
information including title, control numbers, call numbers, and
call number subject classifications. Mainly for testing.
Will be expanded in future versions.
}
\keyword{datasets}
