% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books_serials_etc_sample_data.R
\docType{data}
\name{books_serials_etc_sample}
\alias{books_serials_etc_sample}
\title{Small sample of books, monographs, and serials and their information}
\format{
An object of class \code{"data.table"};
}
\usage{
data(books_serials_etc_sample)
}
\description{
A very small sample of books, serials, VHSs, CDs, and Computer files
and some information including title, control numbers, call numbers, and
call number subject classifications. Somewhat messy/inconsistent
(deliberately) and mainly for testing.
Will be expanded in future versions.
}
\keyword{datasets}
