% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split-map-filter-reduce.R
\name{car}
\alias{car}
\title{Return first element of vector}
\usage{
car(x)
}
\arguments{
\item{x}{A vector}
}
\value{
Returns first element of vector
}
\description{
Takes a vector and returns the first element
Equivalent to Lisp's \code{car} function
}
\details{
Originally for use as a reduction function in \code{split_map_filter_reduce}
}
\examples{
car(c(8, 6, 7, 5, 3, 0, 9))      # 8
mt <- as.data.table(mtcars)
dt_del_cols(mt, "cyl", "disp", "hp")

}
\seealso{
\code{\link{split_map_filter_reduce}}
}
