% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibcodes.R
\name{check_isbn_10_check_digit}
\alias{check_isbn_10_check_digit}
\title{Check the check digit of an ISBN 10}
\usage{
check_isbn_10_check_digit(x, error.is.false = FALSE)
}
\arguments{
\item{x}{A string of or 10 digits or nine digits with terminal "X"}

\item{error.is.false}{return false if error instead of throwing error}
}
\value{
Returns TRUE if check passes, FALSE if not, and NA if NA
}
\description{
Takes a string representation of an ISBN 10 and verifies that check digit
checks out
}
\examples{

check_isbn_10_check_digit("012491540X")  # TRUE

# vectorized
check_isbn_10_check_digit(c("012491540X", "9004037812"))  # TRUE FALSE

}
