% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country_code_crosswalk-data.R
\docType{data}
\name{country_code_crosswalk}
\alias{country_code_crosswalk}
\title{Country code / country crosswalk}
\format{
An object of class \code{"data.table"};
}
\source{
\url{https://www.loc.gov/marc/countries/countries_code.html}
}
\usage{
data(country_code_crosswalk)
}
\description{
A cross-walk between the country code
and it's human readable version
}
\keyword{datasets}
