% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countryxwalk-data.R
\docType{data}
\name{countryxwalk}
\alias{countryxwalk}
\title{Country code / country crosswalk}
\format{
An object of class \code{"data.table"};
}
\usage{
data(countryxwalk)
}
\description{
A cross-walk between the country code
and it's human readable version
}
\keyword{datasets}
