% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dewey_subject_crosswalk-data.R
\docType{data}
\name{dewey_subject_crosswalk}
\alias{dewey_subject_crosswalk}
\title{Dewey Decimal Classification / Subject Description crosswalk}
\format{
An object of class \code{"data.table"};
}
\source{
Edited from \url{https://www.oclc.org/content/dam/oclc/dewey/ddc23-summaries.pdf}
}
\usage{
data(dewey_subject_crosswalk)
}
\description{
A cross-walk between the Dewey Decimal Classification code
and it's human readable subject description
}
\keyword{datasets}
