% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{dt_del_cols}
\alias{dt_del_cols}
\title{Delete columns in a data.table}
\usage{
dt_del_cols(DT, ...)
}
\arguments{
\item{DT}{A data.table}

\item{...}{arbitrary number of column names in quotes}
}
\value{
Returns data.table with those columns removed
}
\description{
Takes a data.table and a quoted sequence of column names
and removes the specified column names from the data.tablee
}
\examples{
mt <- as.data.table(mtcars)
dt_del_cols(mt, "cyl", "disp", "hp")

}
