% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{dt_keep_cols}
\alias{dt_keep_cols}
\title{Keep columns in a data.table}
\usage{
dt_keep_cols(DT, ...)
}
\arguments{
\item{DT}{A data.table}

\item{...}{arbitrary number of column names in quotes}
}
\value{
Returns data.table with only those columns
}
\description{
Takes a data.table and a quoted sequence of column names
and removes all columns but the ones specified
}
\examples{
library(data.table)
mt <- as.data.table(mtcars)
dt_keep_cols(mt, "mpg", "am", "gear", "carb")

}
