% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/language_code_crosswalk-data.R
\docType{data}
\name{language_code_crosswalk}
\alias{language_code_crosswalk}
\title{language code / language crosswalk}
\format{
An object of class \code{"data.table"};
}
\source{
\url{https://www.loc.gov/marc/languages/language_code.html}
}
\usage{
data(language_code_crosswalk)
}
\description{
A cross-walk between the language code
and it's human readable version
}
\keyword{datasets}
