% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/langxwalk-data.R
\docType{data}
\name{langxwalk}
\alias{langxwalk}
\title{language code / language crosswalk}
\format{
An object of class \code{"data.table"};
}
\source{
\url{https://www.loc.gov/marc/languages/language_code.html}
}
\usage{
data(langxwalk)
}
\description{
A cross-walk between the language code
and it's human readable version
}
\keyword{datasets}
