% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lc_first_letter_subject-data.R
\docType{data}
\name{lc_first_letter_subject}
\alias{lc_first_letter_subject}
\title{LC Call Number first letter/subject crosswalk}
\format{
An object of class \code{"data.table"};
}
\usage{
data(lc_first_letter_subject)
}
\description{
A cross-walk between the first letter of a LC Call Number
and it's subject
}
\keyword{datasets}
