% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lc_subject_classification-data.R
\docType{data}
\name{lc_subject_classification}
\alias{lc_subject_classification}
\title{LC Call Subject Code Classification / Classification name crosswalk}
\format{
An object of class \code{"data.table"};
}
\source{
\url{https://www.loc.gov/catdir/cpso/lcco/}
}
\usage{
data(lc_subject_classification)
}
\description{
A cross-walk between the LC Subject classification and
and it's human readable name (first letter of LC Call Number)
}
\keyword{datasets}
