% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lc_subject_subclassification-data.R
\docType{data}
\name{lc_subject_subclassification}
\alias{lc_subject_subclassification}
\title{LC Subject Subclassification / Subclassification name crosswalk}
\format{
An object of class \code{"data.table"};
}
\source{
\url{https://www.loc.gov/catdir/cpso/lcco/}
}
\usage{
data(lc_subject_subclassification)
}
\description{
A cross-walk between the LC Subject subclassification and
and it's human readable name (all letters in an LC Call Number)
}
\keyword{datasets}
