% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lc_two_letter_subject-data.R
\docType{data}
\name{lc_two_letter_subject}
\alias{lc_two_letter_subject}
\title{LC Call Number second letter/subject crosswalk}
\format{
An object of class \code{"data.table"};
}
\source{
\url{https://www.loc.gov/catdir/cpso/lcco/}
}
\usage{
data(lc_two_letter_subject)
}
\description{
A cross-walk between the first two letters of a LC Call Number
and it's subject
}
\keyword{datasets}
