% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibcodes.R
\name{normalize_isbn_10}
\alias{normalize_isbn_10}
\title{Attempt to enforce validity and canonical form to ISBN 10}
\usage{
normalize_isbn_10(x, aggresive = TRUE)
}
\arguments{
\item{x}{A string of or 10 digits or nine digits with terminal "X"}

\item{aggresive}{A logical indicating whether aggresive measures
should be taken to try to get the "ISBN 10"
into a valid form. See "Details" for more info}
}
\value{
Returns TRUE if checks pass, FALSE if not, and NA if NA
}
\description{
Takes a string representation of an ISBN 10. Strips all non-digit
or "X" characters and checks if it is valid (whether the
check digit works out, etc). User can specify whether "aggresive"
measures should be taken to salvage the malformed ISBN 10 string.
}
\details{
If \code{aggresive} is TRUE, aggresive measures are taken to
try to salvage the malformed ISBN 10 string. If the ISBN 10, for example,
is 9 digits, and either adding an "X" to the end, or leading "0"s fix it,
this function will return the salvaged ISBN 10
}
\examples{

is_valid_isbn_10("012491540X")  # TRUE

# vectorized
is_valid_isbn_10(c("012491540X", "9004037812"))  # TRUE FALSE
is_valid_isbn_10(c("012491540X", "hubo un tiempo"))  # TRUE FALSE

}
