% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibcodes.R
\name{normalize_lccn}
\alias{normalize_lccn}
\title{Attempt to enforce validity and canonical form to LCCN}
\usage{
normalize_lccn(x, year.cutoff = NA, include.revisions = FALSE,
  pad.char = "#")
}
\arguments{
\item{x}{A string}

\item{year.cutoff}{STUB}

\item{include.revisions}{STUB}

\item{pad.char}{STUB}
}
\value{
Returns valid LCCN if possible, NA if not
}
\description{
Takes a string representation of an LCCN. THIS IS A STUB
}
\details{
THIS IS A STUB
}
\examples{

# THIS IS A STUB
print("hi")

}
