% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worldcat-api.R
\name{worldcat_api_search}
\alias{worldcat_api_search}
\title{Use the WorldCat Search API}
\usage{
worldcat_api_search(
  sru,
  max_records = 10,
  frbrGrouping = "on",
  start_at = 1,
  wskey = getOption("libbib.wskey", ""),
  more = TRUE,
  print.progress = TRUE,
  debug = FALSE
)
}
\arguments{
\item{sru}{STUB}

\item{max_records}{STUB}

\item{frbrGrouping}{STUB}

\item{start_at}{STUB}

\item{wskey}{STUB}

\item{more}{STUB}

\item{print.progress}{STUB}

\item{debug}{STUB}
}
\value{
STUB
}
\description{
STUB
}
\details{
STUB
}
\examples{

# STUB

}
