% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibcodes.R
\name{normalize_lccn}
\alias{normalize_lccn}
\title{Attempt to enforce validity and canonical form to LCCN}
\usage{
normalize_lccn(x, year.cutoff = NA, include.revisions = FALSE, pad.char = "#")
}
\arguments{
\item{x}{A string (or vector of strings) of LCCNs}

\item{year.cutoff}{If LCCN year is more recent than this year, returns NA}

\item{include.revisions}{STUB}

\item{pad.char}{STUB}
}
\value{
Returns valid LCCN if possible, NA if not
}
\description{
Takes a string representation of an LCCN. Returns a standardized one
}
\details{
THIS IS A STUB
}
\examples{

normalize_lccn("75425165")                   # "###75425165#"

normalize_lccn("2004392586")

# vectorized
normalize_lccn(c("75425165", "2004392586"))
# "###75425165#" "##2004392586"

}
