% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split-map-filter-reduce.R
\name{remove_duplicates_and_nas}
\alias{remove_duplicates_and_nas}
\title{Remove duplicate elements and NAs from a vector}
\usage{
remove_duplicates_and_nas(x)
}
\arguments{
\item{x}{A vector}
}
\value{
Returns vector with duplicates and NAs removed
}
\description{
Takes a vector and returns the same vector without duplicate
elements and without NA values
}
\details{
Can be used as a filtering function in \code{split_map_filter_reduce}
}
\examples{

remove_duplicates_and_nas(c(8, 6, 7, 5, 3, 0, 9, 6, NA, 3))
# 8 6 7 5 3 0 9

remove_duplicates_and_nas(c(NA, NA))
# NA

}
\seealso{
\code{\link{split_map_filter_reduce}}
}
